dont refetch clusters. consolidate into exisitng cluster report
