+++ lib/console/compliance/report.ex
@@ -1,5 +1,5 @@
 defmodule Console.Compliance.Report do
-  alias Console.Compliance.Datasource.{Clusters, Services, Vulnerabilities, ClusterUsers}
+  alias Console.Compliance.Datasource.{Clusters, Services, Vulnerabilities}
   alias Console.Deployments.{Policies, Settings, Policy}
   alias Console.Schema.{User, ComplianceReportGenerator}
 
@@ -17,7 +17,6 @@ defmodule Console.Compliance.Report do
     Zstream.zip([
       Zstream.entry("clusters.csv", Clusters.stream() |> CSV.encode(headers: true)),
       Zstream.entry("services.csv", Services.stream() |> CSV.encode(headers: true)),
-      Zstream.entry("cluster_users.csv", ClusterUsers.stream() |> CSV.encode(headers: true)),
       Zstream.entry("vulnerabilities.csv", Vulnerabilities.stream() |> CSV.encode(headers: true)),
     ])
   end
@@ -26,7 +25,6 @@ defmodule Console.Compliance.Report do
     Zstream.zip([
       Zstream.entry("clusters.json", Clusters.stream() |> Stream.map(&Jason.encode!/1)),
       Zstream.entry("services.json", Services.stream() |> Stream.map(&Jason.encode!/1)),
-      Zstream.entry("cluster_users.json", ClusterUsers.stream() |> Stream.map(&Jason.encode!/1)),
       Zstream.entry("vulnerabilities.json", Vulnerabilities.stream() |> Stream.map(&Jason.encode!/1)),
     ])
   end
